#1212 move the only tests that were getting picked up
